+++ backend/SubManager.Application/DTO/Subscription/SubscriptionsResponseDto.cs
@@ -9,7 +9,7 @@ namespace SubManager.Application.DTO.Subscription
     public class SubscriptionsResponseDto
     {
         public List<SubscriptionDto> Subscriptions { get; set; } = new List<SubscriptionDto>();
-        public float TotalCostMonth { get; set; }
-        public float TotalCostYear { get; set; }
+        public decimal TotalCostMonth { get; set; }
+        public decimal TotalCostYear { get; set; }
     }
 }
